A 64-bit number is a number too: `read_upto("\n", -1)`, `get_size()`
bind-gir writes every GLib integer as `CNumber<…>` but `GType`, which is
an identifier and keeps every bit as a `bigint`. A size, a length, an
offset, a signal handler id is a quantity, and rounds past 2^53 as it
does in GJS. `CNumber` gains the 64-bit C types.

The program-side representation follows TypeScript's type rather than
C's width: an optional brand over `number` is a number however wide its
C type (`native::over_bigint`). It was a `bigint` for any 64-bit brand,
which the checker never saw -- `objc:types`' Swift `Int` too -- and a
Promise form whose `_finish` returns a `gsize` was refused as "settling
with a `bigint`". Four such forms now exist (121 in all), and a `_finish`
returning a number settles through the promise.

gtk-gir drops `BigInt(text.length) as c_size_t`. Test: a `size_t`
through and back as a number, and a promise of one, which with the
representation fix undone is refused as before.

+++ examples/interop/gtk-gir/src/main.ts
@@ -80,7 +80,7 @@ import {
   KeyFileFlags,
   type GError,
 } from "c:GLib-2.0";
-import type { CNumber, c_size_t } from "c:types";
+import type { CNumber } from "c:types";
 import { local } from "c:memory";
 import { gir_emit, gir_log } from "c:gir-shim";
 import { asGtkBox, asGtkButton, asGtkLabel } from "../types/gir/Gtk-4.0.values.ts";
@@ -105,7 +105,7 @@ function outParameters(): void {
   const keys = g_key_file_new();
   const error = local<GError | null>();
   const text = "[a]\nk=5\n";
-  const loaded = g_key_file_load_from_data(keys, text, BigInt(text.length) as c_size_t, KeyFileFlags.NONE, error);
+  const loaded = g_key_file_load_from_data(keys, text, text.length, KeyFileFlags.NONE, error);
   const k = g_key_file_get_integer(keys, "a", "k", error);
   gir_log("keyfile " + String(loaded) + " " + String(k) + " " + (error[0] === null ? "no-error" : "error"));
   const missing = g_key_file_get_integer(keys, "a", "absent", error);
